Fix a bunch of warnings from casts

+++ cxx4/ncDim.cpp
@@ -91,7 +91,7 @@ bool NcDim::isUnlimited() const
   ncCheck(nc_inq_unlimdims(groupId,&numlimdims,unlimdimidsp),__FILE__,__LINE__);
   if (numlimdims){
 	  // get all the unlimited dimension ids in this group
-	  vector<int> unlimdimid(numlimdims);
+	  vector<int> unlimdimid(static_cast<std::size_t>(numlimdims));
 	  ncCheck(nc_inq_unlimdims(groupId,&numlimdims,&unlimdimid[0]),__FILE__,__LINE__);
 	  vector<int>::iterator it;
 	  // now look to see if this dimension is unlimited
